requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
Async-OAuthlib==0.0.9
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
ffmpeg-python==0.2.0
frozenlist==1.3.1
future==0.18.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio-ffmpeg==0.4.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
nextcord==2.2.0
nextcord-ext-ipc==2.2.1
numpy==1.23.3
oauthlib==3.2.1
priority==2.0.0
pycord==0.1.1
PyJWT==2.5.0
python-dotenv==0.21.0
quart==0.18.0
Quart-Discord==2.1.4
requests==2.28.1
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.1
